backend/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
pydantic==2.10.6
pydantic-settings==2.7.0
python-dotenv==1.0.1
httpx==0.28.1
python-multipart==0.0.20
tldextract==5.1.3
dnspython==2.7.0
whois==0.9.27
requests==2.32.3
networkx==3.4.2
openai>=1.40.0
replicate>=0.30.0
apify-client>=1.8.0
langchain>=0.2.16
langchain-openai>=0.1.25
langchain-community>=0.2.16
chromadb>=0.5.15
numpy>=1.26.0
scikit-learn>=1.5.0
reportlab>=4.2.0
python-dateutil>=2.9.0
websockets>=12.0
aiofiles>=24.1.0
